feat: the long-press menu does its job, and a group is called a group
Four Wave 0 packages on the thread's actions and words.

chat-message-actions-do-the-job. The long-press menu grew Share and live
links, split Remove from Report, and lets a visitor report. URLs in a bubble
were dead grey text, and section 7 rule 4 means a chat is the only place one
can arrive: they are now pressable spans (nested Text, not the Android-only
dataDetectorType), underlined and recoloured because on your own accentDeep
bubble the accent has no contrast. Share runs after the menu closes plus
SHEET_SETTLE_MS, per the presentation trap. A moderator gets Remove with its
own confirmation; Report goes to every reader, with a signed-out visitor
handed the SignUpGate instead of a dead route.

chat-anchor-speaks-to-the-accepter. The thread footer told the accepter the
hello started from the SENDER's profile - backwards. footerAnchor() now
picks the second person for the accepter ("Started from your photo") and
keeps the third for the sender, tested for the property that matters: the
two sides of one chat never read the same sentence.

chat-getting-a-second-person-in. A brand-new group dropped its creator into
an empty thread with "Go first. One line is plenty." - advice for a room
with people in it, not a group of one. Under two members it now says "Your
group is ready." with an Invite people button into group settings. The
share message leads with what Samewhere is and ends with the code, and the
paste-parser test's fixture is the exact new message, so the parser is
always tested against what the app actually sends. The spec's
normalizeInviteCode was not built: inviteTokenFrom already does it.

chat-one-word-per-thing. One vocabulary: a traveler-made thing is a GROUP,
a business's is a room, "chat" is only ever a one-to-one. SPEAKING_OPTIONS
deduplicated into one module with "Only people I pick"; member rows read
"In town until {date}" (typed dates, never presence); inbox dates past a
week are "Mar 4", unambiguous.

Review pass (two lenses) found six things, all fixed here. The sweep had
stopped at an e2e assertion: "Chat is active until" and the four "This chat
closed" strings still called a group a chat on the exact screens the
package covers - renamed everywhere including new-group.tsx and the Maestro
step that pins the label. A count on the business dashboard kept the
retired bare "here" suffix; the moderator's Remove confirmation said
"group" inside a business room. And the link spans had eaten the long-press:
a pressable text fragment claims the touch responder, so holding a message
that is entirely a URL - the classic spam shape, the exact message that
most needs Report - could never open the menu and LAUNCHED the link on
release. The spans now arm the same hoisted openMenu handler, and VoiceOver
gets the URL as a rotor action, since the collapsed bubble hides spans from
it entirely. The admin row's null-departure copy became "No leave date",
which is true for the person it describes.

Pending: simulator re-shoots 22, 23, 24, 25, 26, 27, 27a; device checks on
span long-press and the VoiceOver action.

Gate: typecheck, lint, format clean; 587 tests across 57 suites.

+++ src/features/chat/separators.ts
@@ -48,7 +48,11 @@ export function separatorFor(current: MessageRow, older: MessageRow | undefined)
  * a row can spare.
  */
 const WEEKDAY = new Intl.DateTimeFormat('en', { weekday: 'short' });
-const SHORT_DATE = new Intl.DateTimeFormat('en', { month: 'numeric', day: 'numeric' });
+// 'Mar 4', never '3/4': a numeric date means March 4 to an American and
+// 3 April to nearly everyone else this app is for. Unambiguous in every
+// Latin-script locale, and one or two characters wider in a column already
+// sized for 'Yesterday'.
+const SHORT_DATE = new Intl.DateTimeFormat('en', { month: 'short', day: 'numeric' });
 
 export function rowTimestamp(iso: string | null, now: Date = new Date()): string {
   if (!iso) {
